1. 新增對imanhua的支援。 2. 新增對veryim的支援。 3. 翻新178的解析方式，直接轉碼而非參照字典檔。

+++ jcomicdownloader/module/RunModule.java
@@ -147,8 +147,16 @@ public class RunModule {
             }
 
             // 若已存在同檔名壓縮檔，則不解析下載網址
-            if ( !existZipFile( parse.getDownloadDirectory() ) ) {
-                parse.parseComicURL(); // EH在此已經開始下載了。
+            if ( parse.siteID == Site.CK_NOVEL ||
+                 parse.siteID == Site.MYBEST ) { // 針對小說
+                if ( !existTextFile( parse.getDownloadDirectory() ) ) {
+                    parse.parseComicURL(); // 下載網頁檔
+                }
+            }
+            else { // 針對漫畫
+                if ( !existZipFile( parse.getDownloadDirectory() ) ) {
+                    parse.parseComicURL(); // EH在此已經開始下載了。
+                }
             }
             String[] urls = parse.getComicURL();
 
@@ -161,24 +169,39 @@ public class RunModule {
                 System.out.println( "Download Directory : " + parse.getDownloadDirectory() );
                 System.out.println( "\nReady to download ...\n" );
 
-                // 如果已經有同檔名壓縮檔存在，就假設已經下載完畢而不下載。
-                if ( !existZipFile( parse.getDownloadDirectory() ) ) {
-                    Common.debugPrintln( "開始下載整集：" );
-                    Common.downloadManyFile( urls, parse.getDownloadDirectory(),
-                        SetUp.getPicFrontName(), "jpg" );
+                // 基本上，下載小說不會用到downloadManyFile()，故若是小說類就無須檢查
+                if ( parse.siteID != Site.CK_NOVEL ||
+                     parse.siteID == Site.MYBEST ) {
+                    // 如果已經有同檔名壓縮檔存在，就假設已經下載完畢而不下載。
+                    if ( !existZipFile( parse.getDownloadDirectory() ) ) {
+                        Common.debugPrintln( "開始下載整集：" );
+                        Common.downloadManyFile( urls, parse.getDownloadDirectory(),
+                            SetUp.getPicFrontName(), "jpg" );
+                    }  
                 }
             }
+            
+             
+                if ( new File( parse.getDownloadDirectory() ).exists() ) // 存在下載圖檔資料夾
+                {
+                    followingWork( parse );
+                }
+                else {
+                    if ( parse.siteID == Site.CK_NOVEL ||
+                         parse.siteID == Site.MYBEST ) { // 下載小說
+                        if (  !existTextFile( parse.getDownloadDirectory() ) ) { // 已有壓縮檔當然不用產生資料夾
+                            Flag.downloadErrorFlag = true; // 發生錯誤
+                            Common.errorReport( "ERROR： 沒有產生" + parse.getDownloadDirectory() );
+                        }
+                    }
+                    else { // 下載漫畫
+                        if (  !existZipFile( parse.getDownloadDirectory() ) ) { // 已有壓縮檔當然不用產生資料夾
+                            Flag.downloadErrorFlag = true; // 發生錯誤
+                            Common.errorReport( "ERROR： 沒有產生" + parse.getDownloadDirectory() );
+                        }
+                    }
+                }
 
-            if ( new File( parse.getDownloadDirectory() ).exists() ) // 存在下載圖檔資料夾
-            {
-                followingWork( parse );
-            }
-            else {
-                if ( !existZipFile( parse.getDownloadDirectory() ) ) { // 已有壓縮檔當然不用產生資料夾
-                    Flag.downloadErrorFlag = true; // 發生錯誤
-                    Common.errorReport( "ERROR： 沒有產生" + parse.getDownloadDirectory() );
-                }
-            }
             Flag.downloadingFlag = false;
         }
     }
@@ -200,7 +223,12 @@ public class RunModule {
             || siteID == Site.EMLAND
             || siteID == Site.MOP
             || siteID == Site.DM5
-            || siteID == Site.IASK ) {
+            || siteID == Site.IASK
+            || siteID == Site.CK_NOVEL  
+            || siteID == Site.MYBEST
+            || siteID == Site.IMANHUA
+            || siteID == Site.VERYIM
+      ) {
             return true;
         }
         else {
@@ -226,33 +254,60 @@ public class RunModule {
             return false;
         }
     }
+    
+    public boolean existTextFile( String downloadPicDirectory ) {
+        String file;
+        if ( downloadPicDirectory.lastIndexOf( Common.getSlash() ) == downloadPicDirectory.length() - 1 ) {
+            file = downloadPicDirectory.substring( 0, downloadPicDirectory.length() - 1 ) + ".txt";
+        }
+        else {
+            file = downloadPicDirectory + ".txt";
+        }
+
+        //Common.debugPrintln( "將處理的文件檔名稱：" + file );
+        if ( new File( file ).exists() && new File( file ).length() > 1024 ) {
+            Common.debugPrintln( file + "已經存在!" );
+            return true;
+        }
+        else {
+            return false;
+        }
+    }
 
     public synchronized void followingWork( ParseOnlineComicSite parse ) { // compress or not & delete or not
         if ( Run.isAlive ) {
             File downloadPath = new File( parse.getDownloadDirectory() );
 
-            if ( SetUp.getAutoCompress() ) { // compress to zip file or not
-
-                File zipFile = new File( downloadPath.getParent() + "/"
-                    + parse.getWholeTitle() + "." + SetUp.getCompressFormat() );
-
-                if ( downloadPath.list().length < 1 ) {
-                    Common.debugPrintln( "不產生壓縮檔（" + downloadPath.getAbsolutePath() + "資料夾內沒有任何檔案）" );
-                }
-                else if ( zipFile.exists() && zipFile.length() > 1024 ) {
-                    Common.debugPrintln( "不產生壓縮檔（" + zipFile.getAbsolutePath() + "已存在）" );
-                }
-                else {
-                    Common.compress( downloadPath, zipFile );
-                    System.out.println( zipFile.getAbsolutePath() + " made!" );
-                }
-
+            if ( parse.siteID == Site.CK_NOVEL || 
+                 parse.siteID == Site.MYBEST ) {
+                Common.debugPrintln( "小說類型不產生壓縮檔！" );
             }
+            else { 
+                if ( SetUp.getAutoCompress() ) { // compress to zip file or not
 
-            if ( SetUp.getDeleteOriginalPic() ) { // delete the whole folder or not
-                Common.deleteFolder( parse.getDownloadDirectory() );
+                    File zipFile = new File( downloadPath.getParent() + "/"
+                        + parse.getWholeTitle() + "." + SetUp.getCompressFormat() );
 
-                System.out.println( parse.getDownloadDirectory() + " deletion!" );
+                    if ( downloadPath.list().length < 1 ) {
+                        Common.debugPrintln( "不產生壓縮檔（" + downloadPath.getAbsolutePath() + "資料夾內沒有任何檔案）" );
+                    }
+                    else if ( zipFile.exists() && zipFile.length() > 1024 ) {
+                        Common.debugPrintln( "不產生壓縮檔（" + zipFile.getAbsolutePath() + "已存在）" );
+                    }
+                    else {
+                        Common.compress( downloadPath, zipFile );
+                        System.out.println( zipFile.getAbsolutePath() + " made!" );
+                    }
+
+                }
+                
+                if ( SetUp.getDeleteOriginalPic() ) { // delete the whole folder or not
+                    Common.deleteFolder( parse.getDownloadDirectory() );
+
+                    System.out.println( parse.getDownloadDirectory() + " deletion!" );
+                }
+
+                
             }
         }
     }
